Address all warnings with -Wall flags in GCC 9.3

+++ src/backend/cuda/copy.cpp
@@ -63,7 +63,7 @@ Array<T> copyArray(const Array<T> &src) {
 template<typename inType, typename outType>
 Array<outType> padArray(Array<inType> const &in, dim4 const &dims,
                         outType default_value, double factor) {
-    ARG_ASSERT(1, (in.ndims() == (size_t)dims.ndims()));
+    ARG_ASSERT(1, (in.ndims() == dims.ndims()));
     Array<outType> ret = createEmptyArray<outType>(dims);
     kernel::copy<inType, outType>(ret, in, in.ndims(), default_value, factor);
     return ret;
@@ -100,7 +100,7 @@ template<typename inType, typename outType>
 void copyArray(Array<outType> &out, Array<inType> const &in) {
     static_assert(!(is_complex<inType>::value && !is_complex<outType>::value),
                   "Cannot copy from complex value to a non complex value");
-    ARG_ASSERT(1, (in.ndims() == (size_t)out.dims().ndims()));
+    ARG_ASSERT(1, (in.ndims() == out.dims().ndims()));
     copyWrapper<inType, outType> copyFn;
     copyFn(out, in);
 }
